Fix suggested changes - Remove redundant admin policy - Remove set_admin as it's unused - Move utils back - Fix test fixtures and policy count
